refactor: Implement useRichTextEditor hook and extract utility functions for improved code organization and readability
