+++ src/constants/richTextEditorConstants.js
@@ -1,0 +1,20 @@
+export const DEFAULT_BUTTON_STATES = {
+  bold: false,
+  italic: false,
+  underline: false,
+  strikeThrough: false,
+  insertUnorderedList: false,
+  insertOrderedList: false,
+  heading: false,
+};
+
+export const NAVIGATION_KEYS = [
+  "ArrowUp",
+  "ArrowDown",
+  "ArrowLeft",
+  "ArrowRight",
+  "Home",
+  "End",
+  "PageUp",
+  "PageDown",
+];
